requirements.txt
fastapi
uvicorn
langchain
langgraph
python-dotenv
httpx
requests
Pillow
google-genai